gitlab-ci: build and install LittleCMS version 2.16
When using some features from LittleCMS in our CI, we are seeing some
crashes on the address sanitizer. Bumping the LittleCMS version fixes
that. So build and install a more recent version of LittleCMS on our CI.

We chose version 2.16 because it introduces the function
cmsGetToneCurveSegment(). We already make extensive use of that in our
codebase, so it is a good idea to have that on our CI as well.

Now color-curve-segments.c will start to get build on the CI, as
HAVE_CMS_GET_TONE_CURVE_SEGMENT will be true. So we also fix a minor
issue in which we were comparing int with uint in this file, what was
caught after experimenting bumping the LittleCMS version.

+++ libweston/color-lcms/color-curve-segments.c
@@ -301,7 +301,7 @@ are_segments_equal(const cmsCurveSegment *seg_A, const cmsCurveSegment *seg_B)
 	/* Parametric curve. Determine the number of params that we should
 	 * compare. */
 	for (i = 0; i < ARRAY_LENGTH(types); i++) {
-		if (types[i] == abs(seg_A->Type)) {
+		if (types[i] == (uint32_t)abs(seg_A->Type)) {
 			n_params = types_n_params[i];
 			break;
 		}
